kubernetes: allow kubelet to create unlabeled dirs
When a pod volumeMount is used with a subPath specified, kubelet will
first mount the underlying volume and then create the directory for the
subPath if it does not exist. If this is the first time that this volume
has been mounted, the volume has not yet been relabeled and so its
contents will be unlabeled.

Allow kubelet to create unlabeled directories in this situation.

Signed-off-by: Kenton Groombridge <concord@gentoo.org>